requirements.txt
accelerate
bitsandbytes
chainlit
chromadb
ctransformers
faiss-cpu
huggingface_hub
langchain
openai
pandas
PyPDF2
pypdf
sentence_transformers
streamlit
streamlit-chat
tiktoken
torch
camelot-py
langchain_groq
pandasai
pandasql
google-generativeai
plotly